feat: cli
