+++ package.json
@@ -4,10 +4,10 @@
   "description": "This repository is the part of nodejs-assignments https://github.com/AlreadyBored/nodejs-assignments",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "cli": "node ./src/cli/cli.js  --value1 1 --value2 2"
   },
   "repository": {
-    "type": "git",
     "url": "git+https://github.com/AlreadyBored/node-nodejs-basics.git"
   },
   "keywords": ["nodejs", "assignments", "alreadybored"],
