+++ src/cli/cli.js
@@ -1,0 +1,8 @@
+global.RSS_name1 = 10;
+global.RSS_name2 = 20;
+
+import { parseArgs } from "./args.js";
+import { parseEnv } from "./env.js";
+
+parseArgs()
+parseEnv()
